backend/requirements.txt
fastapi==0.109.0
pydantic==1.10.14
uvicorn==0.15.0
starlette==0.35.1
pillow==10.2.0
## The following requirements were added by pip freeze:
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
diffusers==0.25.0
dm-tree==0.1.8
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.2
fsspec==2023.12.2
gast==0.5.4
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.20.2
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
invisible-watermark==0.2.0
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab-widgets==3.0.9
keras==2.15.0
keras-nightly==3.0.4.dev2024012603
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.9
ml-dtypes==0.3.1
mpmath==1.3.0
namex==0.0.7
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic_core==2.16.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.5.0
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.1
scikit-image==0.22.0
scipy==1.12.0
segment-anything==1.0
segment-anything-py==1.0
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
-e git+https://github.com/Stability-AI/stablediffusion.git@cf1d67a6fd5ea1aa600c4df58e5b47da45f6bdbf#egg=stable_diffusion
stack-data==0.6.3
sympy==1.12
tb-nightly==2.16.0a20240127
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-docs @ git+https://github.com/tensorflow/docs@b12399eb2047f534a7474fff58c16b4e3be79886
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf-nightly==2.16.0.dev20240126
tf_keras-nightly==2.16.0.dev2024012710
tf_nightly_intel==2.16.0.dev20240126
tifffile==2023.12.9
tokenizers==0.12.1
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.19.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
zipp==3.17.0